backend/requirements.txt
flask==3.1.2
flask-cors==6.0.2
google-generativeai==0.8.6
python-dotenv==1.0.0
mysql-connector-python==9.5.0
PyJWT==2.10.1

